scraperdb/requirements.txt
requests
beautifulsoup4
lxml
pymongo
python-dotenv
playwright
docling
FastAPI
 pydantic
sentence-transformers
chromadb
FlagEmbedding
google-generativeai


